fix: move App.tsx to components
