+++ src/components/App.tsx
@@ -1,9 +1,9 @@
+import { useColorMode } from "@chakra-ui/react";
 import * as React from "react";
 import browser from "webextension-polyfill";
-import { useColorMode } from "@chakra-ui/react";
-import Page from "./components/Page";
-import { setStoredColorMode, StorageKey } from "./lib/webextension";
-import { LinksProvider } from "./contexts/Links";
+import { LinksProvider } from "../contexts/Links";
+import { setStoredColorMode, StorageKey } from "../lib/webextension";
+import Page from "./Page";
 
 export default function App(): JSX.Element {
   const { colorMode, toggleColorMode } = useColorMode();
