+++ src/index.js
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 import { ChakraProvider, ColorModeScript } from "@chakra-ui/react";
-import App from "./App";
+import App from "./components/App";
 import "./index.css";
 import theme from "./theme";
